Make all Valkyrie forms inherit from ResourceForm
Formerly, `Hyrax::Forms::ResourceForm` contained a bunch of shared
behaviours between objects and filesets which weren’t appropriate
for adminsets or collections. This commit splits those behaviours
out into a new `Hyrax::ContainedInWorksBehavior` module, which
allows all forms to inherit from the same base class.
