+++ app/forms/hyrax/forms/resource_form.rb
@@ -6,13 +6,7 @@ module Hyrax
     # @api public
     #
     # Returns the form class associated with a given model.
-    #
-    # @note The default case assumes that the provided model class is for a
-    #   PCDM object and returns a +Hyrax::Forms::PcdmObjectForm+. This is for
-    #   backwards‐compatibility with existing Hyrax instances. However, a
-    #   different +Hyrax::Forms::ResourceForm+ subclass will be returned in
-    #   some known cases where that is preferable.
-    def self.ResourceForm(model_class)
+    def self.ResourceForm(model_class) # rubocop:disable Metrics/MethodLength, Metrics/PerceivedComplexity
       @resource_forms ||= {}.compare_by_identity
       @resource_forms[model_class] ||=
         if model_class <= Hyrax::AdministrativeSet
@@ -31,20 +25,6 @@ module Hyrax
     #
     # This form wraps +Hyrax::ChangeSet+ in the +HydraEditor::Form+ interface.
     class ResourceForm < Hyrax::ChangeSet # rubocop:disable Metrics/ClassLength
-      ##
-      # @api private
-      InWorksPrepopulator = proc do |_options|
-        self.in_works_ids =
-          if persisted?
-            Hyrax.query_service
-                 .find_inverse_references_by(resource: model, property: :member_ids)
-                 .select(&:work?)
-                 .map(&:id)
-          else
-            []
-          end
-      end
-
       ##
       # @api private
       #
@@ -67,35 +47,8 @@ module Hyrax
       class_attribute :model_class
 
       property :human_readable_type, writable: false
-
-      property :depositor
-
-      property :visibility, default: VisibilityIntention::PRIVATE, populator: :visibility_populator
-
       property :date_modified, readable: false
       property :date_uploaded, readable: false
-      property :agreement_accepted, virtual: true, default: false, prepopulator: proc { |_opts| self.agreement_accepted = !model.new_record }
-
-      collection(:permissions,
-                 virtual: true,
-                 default: [],
-                 form: Hyrax::Forms::Permission,
-                 populator: :permission_populator,
-                 prepopulator: proc { |_opts| self.permissions = Hyrax::AccessControl.for(resource: model).permissions })
-
-      property :embargo, form: Hyrax::Forms::Embargo, populator: :embargo_populator
-      property :lease, form: Hyrax::Forms::Lease, populator: :lease_populator
-
-      # virtual properties for embargo/lease;
-      property :embargo_release_date, virtual: true, prepopulator: proc { |_opts| self.embargo_release_date = model.embargo&.embargo_release_date }
-      property :visibility_after_embargo, virtual: true, prepopulator: proc { |_opts| self.visibility_after_embargo = model.embargo&.visibility_after_embargo }
-      property :visibility_during_embargo, virtual: true, prepopulator: proc { |_opts| self.visibility_during_embargo = model.embargo&.visibility_during_embargo }
-
-      property :lease_expiration_date, virtual: true,  prepopulator: proc { |_opts| self.lease_expiration_date = model.lease&.lease_expiration_date }
-      property :visibility_after_lease, virtual: true, prepopulator: proc { |_opts| self.visibility_after_lease = model.lease&.visibility_after_lease }
-      property :visibility_during_lease, virtual: true, prepopulator: proc { |_opts| self.visibility_during_lease = model.lease&.visibility_during_lease }
-
-      property :in_works_ids, virtual: true, prepopulator: InWorksPrepopulator
 
       # provide a lock token for optimistic locking; we name this `version` for
       # backwards compatibility
@@ -185,37 +138,6 @@ module Hyrax
       end
 
       private
-
-      def embargo_populator(**)
-        self.embargo = Hyrax::EmbargoManager.embargo_for(resource: model)
-      end
-
-      def lease_populator(**)
-        self.lease = Hyrax::LeaseManager.lease_for(resource: model)
-      end
-
-      # https://trailblazer.to/2.1/docs/reform.html#reform-populators-populator-collections
-      def permission_populator(collection:, index:, **)
-        Hyrax::Forms::Permission.new(collection[index])
-      end
-
-      def visibility_populator(fragment:, doc:, **)
-        case fragment
-        when "embargo"
-          self.visibility = doc['visibility_during_embargo']
-
-          doc['embargo'] = doc.slice('visibility_after_embargo',
-                                     'visibility_during_embargo',
-                                     'embargo_release_date')
-        when "lease"
-          self.visibility = doc['visibility_during_lease']
-          doc['lease'] = doc.slice('visibility_after_lease',
-                                     'visibility_during_lease',
-                                     'lease_expiration_date')
-        else
-          self.visibility = fragment
-        end
-      end
 
       def _form_field_definitions
         self.class.definitions
